Create Thermostat: Heating and Cooling Mode
DESCRIPTION: 
Thermostat is a device which we use to regulate the temperature of a closed environments.
There are many types of thermostats, some are manually controlled. Our thermostat has manual and smart control. 
There are two pushbuttons, for down and up regulation respectively. 
Temperature sensor takes the current temperature of our surroundings and displays it on the LCD screen. 
We can manually decrease the temperature up to 16 degrees Celsius, system will operate in the cooling mode, fan will be operational.
Decreasing below 16 degrees will trigger the lower threshold we have set; it will result in turning on the heating element. That is, our system will go into heating mode.
Similarly, we can increase the temperature up to 25 degrees above which the upper threshold will be triggered and the system will go into heating mode. 
When decrementing the current temperature, if we decrement between 20 and 16 degrees, the fan will be working in 3x speed. If we decrement below 16 degrees, system will go into heating mode.
And while incrementing, if we increment between 20 and 25 degrees, then the fan will be working in 1.5x speed. If we increment above 25 degrees, the system will go into cooling mode and work at max speed. 

COMPONENTS REQUIRED: 
1.	Arduino Uno R3 with power supply 
2.	DHT22 temperature and humidity sensor 
3.	Relay x 2
4.	12V Fan 
5.	12V Heating Element (PTC Thermostat heating element)
6.	LCD Display 
7.	Push Button x 2
8.	12V DC voltage adapter
9.	Jumper wires and breadboard

CONNECTIONS:
- Arduino is given an external power supply using jack. The common (ground) and Vout terminals are connected to the breadboard for giving common ports to the engineer. 
- DHT22 temperature sensor is given 5V supply Vin/Vcc and ground. Data terminal is connected to pin 8 of Arduino board. 
- Both relays are given 5V Vcc and ground.
- Relay1 containing the cooling element (12V DC Fan) 
IN terminal of relay is connected to pin 11 (PWM) of Arduino.
Output terminals- COM is connected to the common terminal of the fan and NO is connected to one terminal of the 12V DC adapter and supply from the other terminal of 12V DC adapter is given to the Vcc terminal to the fan.
- Relay2 containing the heating element.
IN terminal of the relay is connected to A6 pin of Arduino.
Output terminals- COM is connected to the ground/common terminal of the 12V DC heating element and the NO port is connected to one terminal of the 12V DC adapter whose other terminal is given as Vcc input to the heating element. 
- Two push buttons (for regulating the temperature) are connected to A2 and A4 respectively. 
- LCD is connected to pins 2 through 7 of the Arduino board.